Move settings into the app shell
